llm_game_agent_template/requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
google-generativeai>=0.3.2 